Add the search card
